+++ src/main/java/com/lalabrand/ecommerce/auth/jwt_token/JwtService.java
@@ -1,5 +1,6 @@
-package com.lalabrand.ecommerce.auth;
+package com.lalabrand.ecommerce.auth.jwt_token;
 
+import com.lalabrand.ecommerce.auth.jwt_token.JwtToken;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
