+++ src/main/java/com/lalabrand/ecommerce/auth/refresh_token/RefreshTokenRepository.java
@@ -1,4 +1,4 @@
-package com.lalabrand.ecommerce.auth;
+package com.lalabrand.ecommerce.auth.refresh_token;
 
 import com.lalabrand.ecommerce.user.User;
 import org.springframework.data.repository.CrudRepository;
